wantHave_backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pillow==11.0.0
channels==4.3.2
daphne==4.2.1
channels_redis==4.3.0
